LF-4383 Extract initial form-setting logic to own hook
